Finished doing some of the backend changes for new systems. There is little change to the front end right now

+++ src/main/java/com/thenexusreborn/hub/scoreboard/HubScoreboard.java
@@ -41,12 +41,12 @@ public class HubScoreboard extends SpigotScoreboardView {
         
         ITeam creditsValue = scoreboard.registerNewTeam(creditsValueName);
         creditsValue.setPrefix(MCUtils.color("&fCredits: "));
-        creditsValue.setSuffix(MCUtils.color("&3" + formatBalance(player.getStatValue("credits"))));
+        creditsValue.setSuffix(MCUtils.color("&3" + formatBalance((double) player.getStatValue("credits"))));
         addEntry(objective, creditsValue, ChatColor.DARK_RED.toString(), 11);
         
         ITeam nexitesValue = scoreboard.registerNewTeam(nexitesValueName);
         nexitesValue.setPrefix(MCUtils.color("&fNexites: "));
-        nexitesValue.setSuffix(MCUtils.color("&9" + formatBalance(player.getStatValue("nexites"))));
+        nexitesValue.setSuffix(MCUtils.color("&9" + formatBalance((double) player.getStatValue("nexites"))));
         addEntry(objective, nexitesValue, ChatColor.AQUA.toString(), 10);
     
         ITeam blank2 = scoreboard.registerNewTeam(blank2Name);
@@ -87,8 +87,8 @@ public class HubScoreboard extends SpigotScoreboardView {
             rankName = player.getRank().name();
         }
         scoreboard.getTeam(this.rankValueName).setPrefix(MCUtils.color(player.getRank().getColor() + rankName));
-        scoreboard.getTeam(this.creditsValueName).setSuffix(MCUtils.color("&3" + formatBalance(player.getStatValue("credits"))));
-        scoreboard.getTeam(this.nexitesValueName).setSuffix(MCUtils.color("&9" + formatBalance(player.getStatValue("nexites"))));
+        scoreboard.getTeam(this.creditsValueName).setSuffix(MCUtils.color("&3" + formatBalance((double) player.getStatValue("credits"))));
+        scoreboard.getTeam(this.nexitesValueName).setSuffix(MCUtils.color("&9" + formatBalance((double) player.getStatValue("nexites"))));
         int onlinePlayers = 0;
         for (Player onlinePlayer : Bukkit.getOnlinePlayers()) {
             NexusPlayer onp = NexusAPI.getApi().getPlayerManager().getNexusPlayer(onlinePlayer.getUniqueId());
